fix added

+++ coffeeshop/src/main/java/com/coffeeshop/Controller/ShopController.java
@@ -3,8 +3,10 @@ package com.coffeeshop.Controller;
 import com.coffeeshop.EntityClasses.QueueEntity;
 import com.coffeeshop.EntityClasses.ShopEntity;
 import com.coffeeshop.EntityClasses.UserEntity;
+import com.coffeeshop.Enum.UserRoleEnum;
 import com.coffeeshop.Model.BaseRestResponse;
 import com.coffeeshop.Model.Request.CreateQueue;
+import com.coffeeshop.Model.Request.CreateUser;
 import com.coffeeshop.Repository.QueueRepository;
 import com.coffeeshop.Repository.UserRepository;
 import com.coffeeshop.Services.ShopService;
@@ -47,13 +49,24 @@ public class ShopController {
                 responseJson = mapper.writeValueAsString(baseResponse);
                 return new ResponseEntity<String>(responseJson, responseHeaders, HttpStatus.UNAUTHORIZED);
             }
-            final ShopEntity entity = shopService.create(shopEntity);
-            baseResponse.setError(false);
-            baseResponse.setData(entity);
-            baseResponse.setMessage("Created Shop Successfully");
-            baseResponse.setCode(201);
-            responseJson = mapper.writeValueAsString(baseResponse);
-            return new ResponseEntity<String>(responseJson, responseHeaders, HttpStatus.OK);
+            String s[] = Authorization.split(" ");
+            final UserEntity byToken = userRepository.findByToken(s[1]);
+            if (byToken.getRole().equals(UserRoleEnum.SHOP_OWNER.getValue())) {
+                final ShopEntity entity = shopService.create(shopEntity);
+                baseResponse.setError(false);
+                baseResponse.setData(entity);
+                baseResponse.setMessage("Created Shop Successfully");
+                baseResponse.setCode(201);
+                responseJson = mapper.writeValueAsString(baseResponse);
+                return new ResponseEntity<String>(responseJson, responseHeaders, HttpStatus.OK);
+            } else {
+                baseResponse.setError(true);
+                baseResponse.setData(null);
+                baseResponse.setMessage("You haven't Permission");
+                responseJson = mapper.writeValueAsString(baseResponse);
+                return new ResponseEntity<String>(responseJson, responseHeaders, HttpStatus.UNAUTHORIZED);
+            }
+
 
         } catch (Exception ex) {
             baseResponse.setError(true);
@@ -104,13 +117,24 @@ public class ShopController {
                 return new ResponseEntity<String>(responseJson, responseHeaders, HttpStatus.UNAUTHORIZED);
             }
 
-            final ShopEntity entity = shopService.Update(shopEntity);
-            baseResponse.setError(false);
-            baseResponse.setData(entity);
-            baseResponse.setMessage("Updated Shop Successfully");
-            baseResponse.setCode(201);
-            responseJson = mapper.writeValueAsString(baseResponse);
-            return new ResponseEntity<String>(responseJson, responseHeaders, HttpStatus.OK);
+            String s[] = Authorization.split(" ");
+            final UserEntity byToken = userRepository.findByToken(s[1]);
+            if (byToken.getRole().equals(UserRoleEnum.SHOP_OWNER.getValue())) {
+                final ShopEntity entity = shopService.Update(shopEntity);
+                baseResponse.setError(false);
+                baseResponse.setData(entity);
+                baseResponse.setMessage("Updated Shop Successfully");
+                baseResponse.setCode(201);
+                responseJson = mapper.writeValueAsString(baseResponse);
+                return new ResponseEntity<String>(responseJson, responseHeaders, HttpStatus.OK);
+            } else {
+                baseResponse.setError(true);
+                baseResponse.setData(null);
+                baseResponse.setMessage("You haven't Permission");
+                responseJson = mapper.writeValueAsString(baseResponse);
+                return new ResponseEntity<String>(responseJson, responseHeaders, HttpStatus.UNAUTHORIZED);
+            }
+
 
         } catch (Exception ex) {
             baseResponse.setError(true);
@@ -145,13 +169,24 @@ public class ShopController {
                 return new ResponseEntity<String>(responseJson, responseHeaders, HttpStatus.UNAUTHORIZED);
             }
 
-            final ShopEntity entity = shopService.Delete(shopEntity);
-            baseResponse.setError(false);
-            baseResponse.setData(entity);
-            baseResponse.setMessage("Deleted Shop Successfully");
-            baseResponse.setCode(201);
-            responseJson = mapper.writeValueAsString(baseResponse);
-            return new ResponseEntity<String>(responseJson, responseHeaders, HttpStatus.OK);
+            String s[] = Authorization.split(" ");
+            final UserEntity byToken = userRepository.findByToken(s[1]);
+            if (byToken.getRole().equals(UserRoleEnum.SHOP_OWNER.getValue())) {
+                final ShopEntity entity = shopService.Delete(shopEntity);
+                baseResponse.setError(false);
+                baseResponse.setData(entity);
+                baseResponse.setMessage("Deleted Shop Successfully");
+                baseResponse.setCode(201);
+                responseJson = mapper.writeValueAsString(baseResponse);
+                return new ResponseEntity<String>(responseJson, responseHeaders, HttpStatus.OK);
+            } else {
+                baseResponse.setError(true);
+                baseResponse.setData(null);
+                baseResponse.setMessage("You haven't Permission");
+                responseJson = mapper.writeValueAsString(baseResponse);
+                return new ResponseEntity<String>(responseJson, responseHeaders, HttpStatus.UNAUTHORIZED);
+            }
+
 
         } catch (Exception ex) {
             baseResponse.setError(true);
@@ -201,6 +236,7 @@ public class ShopController {
             return new ResponseEntity<String>(responseJson, responseHeaders, HttpStatus.OK);
 
         } catch (Exception ex) {
+            ex.printStackTrace();
             baseResponse.setError(false);
             baseResponse.setData(null);
             baseResponse.setMessage("No Shops");
@@ -233,27 +269,39 @@ public class ShopController {
                 responseJson = mapper.writeValueAsString(baseResponse);
                 return new ResponseEntity<String>(responseJson, responseHeaders, HttpStatus.UNAUTHORIZED);
             }
-            QueueEntity queueEntity = new QueueEntity();
-            Optional<ShopEntity> shop = shopService.getShop(CreateQueue.getShopid());
-            if (shop.get() != null) {
-                queueEntity.setShop(shop.get());
-                queueEntity.setMaxsize(CreateQueue.getMax());
+            String s[] = Authorization.split(" ");
+            final UserEntity byToken = userRepository.findByToken(s[1]);
+            if (byToken.getRole().equals(UserRoleEnum.SHOP_OWNER.getValue()) || byToken.getRole().equals(UserRoleEnum.SHOP_OPERATOR.getValue())) {
+
+
+                QueueEntity queueEntity = new QueueEntity();
+                Optional<ShopEntity> shop = shopService.getShop(CreateQueue.getShopid());
+                if (shop.get() != null) {
+                    queueEntity.setShop(shop.get());
+                    queueEntity.setMaxsize(CreateQueue.getMax());
+                } else {
+                    baseResponse.setError(true);
+                    baseResponse.setData(null);
+                    baseResponse.setMessage("Shop is not valid");
+                    baseResponse.setCode(500);
+                    responseJson = mapper.writeValueAsString(baseResponse);
+                    return new ResponseEntity<String>(responseJson, responseHeaders, HttpStatus.INTERNAL_SERVER_ERROR);
+
+                }
+                QueueEntity save = shopService.CreateQueue(queueEntity);
+                baseResponse.setError(false);
+                baseResponse.setData(save);
+                baseResponse.setMessage("Created Queue Successfully");
+                baseResponse.setCode(201);
+                responseJson = mapper.writeValueAsString(baseResponse);
+                return new ResponseEntity<String>(responseJson, responseHeaders, HttpStatus.OK);
             } else {
                 baseResponse.setError(true);
                 baseResponse.setData(null);
-                baseResponse.setMessage("Shop is not valid");
-                baseResponse.setCode(500);
+                baseResponse.setMessage("You haven't Permission");
                 responseJson = mapper.writeValueAsString(baseResponse);
-                return new ResponseEntity<String>(responseJson, responseHeaders, HttpStatus.INTERNAL_SERVER_ERROR);
-
+                return new ResponseEntity<String>(responseJson, responseHeaders, HttpStatus.UNAUTHORIZED);
             }
-            QueueEntity save = shopService.CreateQueue(queueEntity);
-            baseResponse.setError(false);
-            baseResponse.setData(save);
-            baseResponse.setMessage("Created Queue Successfully");
-            baseResponse.setCode(201);
-            responseJson = mapper.writeValueAsString(baseResponse);
-            return new ResponseEntity<String>(responseJson, responseHeaders, HttpStatus.OK);
 
         } catch (Exception ex) {
             baseResponse.setError(true);
@@ -287,27 +335,40 @@ public class ShopController {
                 responseJson = mapper.writeValueAsString(baseResponse);
                 return new ResponseEntity<String>(responseJson, responseHeaders, HttpStatus.UNAUTHORIZED);
             }
-            QueueEntity queueEntity = new QueueEntity();
-            Optional<ShopEntity> shop = shopService.getShop(CreateQueue.getShopid());
-            if (shop.get() != null) {
-                queueEntity.setShop(shop.get());
-                queueEntity.setMaxsize(CreateQueue.getMax());
+            String s[] = Authorization.split(" ");
+            final UserEntity byToken = userRepository.findByToken(s[1]);
+            if (byToken.getRole().equals(UserRoleEnum.SHOP_OWNER.getValue()) || byToken.getRole().equals(UserRoleEnum.SHOP_OPERATOR.getValue())) {
+
+
+                QueueEntity queueEntity = new QueueEntity();
+                Optional<ShopEntity> shop = shopService.getShop(CreateQueue.getShopid());
+                if (shop.get() != null) {
+                    queueEntity.setId(CreateQueue.getId());
+                    queueEntity.setShop(shop.get());
+                    queueEntity.setMaxsize(CreateQueue.getMax());
+                } else {
+                    baseResponse.setError(true);
+                    baseResponse.setData(null);
+                    baseResponse.setMessage("Shop is not valid");
+                    baseResponse.setCode(500);
+                    responseJson = mapper.writeValueAsString(baseResponse);
+                    return new ResponseEntity<String>(responseJson, responseHeaders, HttpStatus.INTERNAL_SERVER_ERROR);
+
+                }
+                QueueEntity save = shopService.UpdateQueue(queueEntity);
+                baseResponse.setError(false);
+                baseResponse.setData(save);
+                baseResponse.setMessage("Update Queue Successfully");
+                baseResponse.setCode(201);
+                responseJson = mapper.writeValueAsString(baseResponse);
+                return new ResponseEntity<String>(responseJson, responseHeaders, HttpStatus.OK);
             } else {
                 baseResponse.setError(true);
                 baseResponse.setData(null);
-                baseResponse.setMessage("Shop is not valid");
-                baseResponse.setCode(500);
+                baseResponse.setMessage("You haven't Permission");
                 responseJson = mapper.writeValueAsString(baseResponse);
-                return new ResponseEntity<String>(responseJson, responseHeaders, HttpStatus.INTERNAL_SERVER_ERROR);
-
+                return new ResponseEntity<String>(responseJson, responseHeaders, HttpStatus.UNAUTHORIZED);
             }
-            QueueEntity save = shopService.UpdateQueue(queueEntity);
-            baseResponse.setError(false);
-            baseResponse.setData(save);
-            baseResponse.setMessage("Update Queue Successfully");
-            baseResponse.setCode(201);
-            responseJson = mapper.writeValueAsString(baseResponse);
-            return new ResponseEntity<String>(responseJson, responseHeaders, HttpStatus.OK);
 
         } catch (Exception ex) {
             baseResponse.setError(true);
@@ -341,27 +402,109 @@ public class ShopController {
                 responseJson = mapper.writeValueAsString(baseResponse);
                 return new ResponseEntity<String>(responseJson, responseHeaders, HttpStatus.UNAUTHORIZED);
             }
-            QueueEntity queueEntity = new QueueEntity();
-            Optional<ShopEntity> shop = shopService.getShop(CreateQueue.getShopid());
-            if (shop.get() != null) {
-                queueEntity.setShop(shop.get());
-                queueEntity.setMaxsize(CreateQueue.getMax());
+            String s[] = Authorization.split(" ");
+            final UserEntity byToken = userRepository.findByToken(s[1]);
+            if (byToken.getRole().equals(UserRoleEnum.SHOP_OWNER.getValue()) || byToken.getRole().equals(UserRoleEnum.SHOP_OPERATOR.getValue())) {
+
+                QueueEntity queueEntity = new QueueEntity();
+                Optional<ShopEntity> shop = shopService.getShop(CreateQueue.getShopid());
+                if (shop.get() != null) {
+                    queueEntity.setShop(shop.get());
+                    queueEntity.setMaxsize(CreateQueue.getMax());
+                    queueEntity.setId(CreateQueue.getId());
+                } else {
+                    baseResponse.setError(true);
+                    baseResponse.setData(null);
+                    baseResponse.setMessage("Shop is not valid");
+                    baseResponse.setCode(500);
+                    responseJson = mapper.writeValueAsString(baseResponse);
+                    return new ResponseEntity<String>(responseJson, responseHeaders, HttpStatus.INTERNAL_SERVER_ERROR);
+
+                }
+                QueueEntity save = shopService.DeleteQueue(queueEntity);
+                baseResponse.setError(false);
+                baseResponse.setData(save);
+                baseResponse.setMessage("Delete Queue Successfully");
+                baseResponse.setCode(201);
+                responseJson = mapper.writeValueAsString(baseResponse);
+                return new ResponseEntity<String>(responseJson, responseHeaders, HttpStatus.OK);
             } else {
                 baseResponse.setError(true);
                 baseResponse.setData(null);
-                baseResponse.setMessage("Shop is not valid");
-                baseResponse.setCode(500);
+                baseResponse.setMessage("You haven't Permission");
                 responseJson = mapper.writeValueAsString(baseResponse);
-                return new ResponseEntity<String>(responseJson, responseHeaders, HttpStatus.INTERNAL_SERVER_ERROR);
-
+                return new ResponseEntity<String>(responseJson, responseHeaders, HttpStatus.UNAUTHORIZED);
             }
-            QueueEntity save = shopService.DeleteQueue(queueEntity);
-            baseResponse.setError(false);
-            baseResponse.setData(save);
-            baseResponse.setMessage("Delete Queue Successfully");
-            baseResponse.setCode(201);
-            responseJson = mapper.writeValueAsString(baseResponse);
-            return new ResponseEntity<String>(responseJson, responseHeaders, HttpStatus.OK);
+
+        } catch (Exception ex) {
+            baseResponse.setError(true);
+            baseResponse.setData(null);
+            baseResponse.setMessage(ex.getMessage().toString());
+            baseResponse.setCode(500);
+            try {
+                responseJson = mapper.writeValueAsString(baseResponse);
+            } catch (JsonProcessingException e) {
+                e.printStackTrace();
+            }
+            return new ResponseEntity<String>(responseJson, responseHeaders, HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+    }
+
+    @PostMapping(value = "/create/subuser")
+    public @ResponseBody
+    ResponseEntity<String> createSubUser(@RequestBody CreateUser createUser,
+                                         @RequestHeader("Authorization") String Authorization) {
+        HttpHeaders responseHeaders = new HttpHeaders();
+        responseHeaders.setContentType(MediaType.APPLICATION_JSON);
+        ObjectMapper mapper = new ObjectMapper();
+        BaseRestResponse baseResponse = new BaseRestResponse();
+        String responseJson = "";
+        try {
+            boolean Auth = validateToken(Authorization);
+            if (!Auth) {
+                baseResponse.setError(true);
+                baseResponse.setData(null);
+                baseResponse.setMessage("Authentication Fail");
+                responseJson = mapper.writeValueAsString(baseResponse);
+                return new ResponseEntity<String>(responseJson, responseHeaders, HttpStatus.UNAUTHORIZED);
+            }
+            String s[] = Authorization.split(" ");
+            final UserEntity byToken = userRepository.findByToken(s[1]);
+            if (byToken.getRole().equals(UserRoleEnum.SHOP_OWNER.getValue())) {
+
+                UserEntity userEntity = new UserEntity();
+                final Optional<ShopEntity> shop = shopService.getShop(createUser.getShopid());
+                if (shop.get() != null) {
+                    userEntity.setShop(shop.get());
+                    userEntity.setUserName(createUser.getUserName());
+                    userEntity.setRole(createUser.getRole());
+                    userEntity.setPassword(createUser.getPassword());
+                    final UserEntity userEntity1 = shopService.CreateUser(userEntity);
+
+                    baseResponse.setError(false);
+                    baseResponse.setData(userEntity1);
+                    baseResponse.setMessage("User Create Successfully");
+                    baseResponse.setCode(201);
+                    responseJson = mapper.writeValueAsString(baseResponse);
+                    return new ResponseEntity<String>(responseJson, responseHeaders, HttpStatus.OK);
+                } else {
+                    baseResponse.setError(true);
+                    baseResponse.setData(null);
+                    baseResponse.setMessage("Shop is not valid");
+                    baseResponse.setCode(500);
+                    responseJson = mapper.writeValueAsString(baseResponse);
+                    return new ResponseEntity<String>(responseJson, responseHeaders, HttpStatus.INTERNAL_SERVER_ERROR);
+
+                }
+
+            } else {
+                baseResponse.setError(true);
+                baseResponse.setData(null);
+                baseResponse.setMessage("You Havent permission");
+                responseJson = mapper.writeValueAsString(baseResponse);
+                return new ResponseEntity<String>(responseJson, responseHeaders, HttpStatus.UNAUTHORIZED);
+            }
+
 
         } catch (Exception ex) {
             baseResponse.setError(true);
